services/ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
httpx>=0.27.0
boto3>=1.35.0
rembg>=2.0.57
Pillow>=11.0.0
numpy>=2.1.0
structlog>=24.4.0
redis>=5.1.1
